Minor corrections
