+++ src/adapter/graphql/tests/tests/mod.rs
@@ -18,6 +18,7 @@ mod test_gql_dataset_flow_runs;
 mod test_gql_datasets;
 mod test_gql_metadata;
 mod test_gql_metadata_chain;
+mod test_gql_remote_statuses;
 mod test_gql_search;
 mod test_guards;
 mod test_update_schema;
