+++ src/e2e/app/cli/repo-tests/src/commands/test_delete_command.rs
@@ -173,6 +173,8 @@ pub async fn test_delete_dataset_all(kamu: KamuCliPuppet) {
     }
 }
 
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 pub async fn test_delete_warning(mut kamu_node_api_client: KamuApiServerClient) {
     let kamu: KamuCliPuppet = KamuCliPuppet::new_workspace_tmp(true).await;
